Replace Clock trait with explicit time injection
Remove the Clock trait, SystemClock, and FakeClock entirely. Store now
accepts time as a parameter (now: u64) rather than holding an Arc<dyn Clock>
and pulling time internally. This makes Store a pure data structure with no
time dependency — tests pass literal timestamps with no mocking infrastructure.

Also fixes TimeT from usize (platform-dependent width) to u64, which avoids
a ~49-day overflow on 32-bit systems.

+++ src/infra/handlers.rs
@@ -29,7 +29,7 @@ impl TimerHandler {
 
 #[derive(Deserialize)]
 struct RequestPayload {
-    interval_ms: usize,
+    interval_ms: u64,
 }
 
 #[async_trait]
